requirements.txt
astroid==2.0.4
awscli==1.15.65
botocore==1.10.64
certifi==2018.1.18
chardet==3.0.4
Click==7.0
colorama==0.3.9
Deprecated==1.2.4
docutils==0.14
Flask==1.0.2
GitHub-Flask==3.2.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pipenv==11.8.0
pyasn1==0.4.4
PyGithub==1.43.3
PyJWT==1.6.4
pylint==2.1.1
python-dateutil==2.7.3
PyYAML==3.13
requests==2.20.1
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
virtualenv==15.1.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
wrapt==1.10.11
